Database_Connection/mysqlDatabase_apiMenu_based/requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
requests==2.32.3
robotframework==7.0.1
robotframework-requests==0.9.7
pytest==8.3.4